feat: Add Rock editor base application
