+++ ore/ore/engine/core/application.cpp
