+++ ore/ore/engine/core/scene_loader.cpp
