+++ ore/ore/engine/renderer/node.cpp
