+++ ore/ore/engine/renderer/renderable.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <memory>
-#include <ore/libs/glad/glad.h>
+#include "ore/libs/glad/glad.h"
 #include "ore/engine/renderer/shader.hpp"
 #include "ore/engine/renderer/world_node.hpp"
 #include "ore/engine/renderer/camera.hpp"
