+++ ore/ore/engine/renderer/scene.cpp
