+++ ore/ore/engine/renderer/scene.hpp
