+++ ore/ore/engine/renderer/world_node.hpp
