+++ ore/ore/libs/glad/glad.c
